chore: move around pinpoint related functions
