+++ packages/amplify-category-notifications/src/plugin-client-api-analytics.ts
@@ -97,3 +97,9 @@ export const invokeAnalyticsPinpointHasInAppMessagingPolicy = async (context: $T
  */
 export const invokeAnalyticsMigrations = async (context: $TSContext): Promise<void> =>
   context.amplify.invokePluginMethod(context, 'analytics', undefined, 'analyticsPluginAPIMigrations', [context]);
+
+/**
+ * invoke analytics plugin migrations
+ */
+export const invokeAnalyticsGetPinpointRegionMapping = async (context: $TSContext): Promise<Record<string, string>> =>
+  context.amplify.invokePluginMethod(context, 'analytics', undefined, 'analyticsGetPinpointRegionMapping', []);
